kernel/{kompile,compile,backend}: always emit allRules.txt, not only when --coverage
